+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 API changes:
 - Add particle batch `invert` field, to spawn particles with reverse motion
 - Add particle batch `pre_spawn_travel` field, to offset particle position relative to motion
+- Add `nature_spark_mini` particle
 
 Functional changes:
 - Update Italian translation, thanks to Zano1999